ScienceProjectTemplate

This is the mandatory index.md file that is the homepage of the documentation.

Here the documentation string of a function is expanded:

dummy_project_function — Function

dummy_project_function(x, y) → z

Dummy function for illustration purposes. Performs operation:

\[z = x + y\]

We can link to any subsection name with Subsection name.

We can link to documentation strings with dummy_project_function.

Subsection name

Test.